Clean up file structure, remove console logs, remove redux logger
