+++ front-end/src/assets/logos/chat_at_table.png
