+++ front-end/src/assets/logos/employee_scheduler2.png
