+++ front-end/src/assets/logos/working_at_comp.png
