+++ front-end/src/assets/logos/working_solo.png
