+++ front-end/src/components/Organisms/BillingStoreCheckout.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Elements } from "react-stripe-elements";
 
-import InjectedCheckoutForm from "./CheckoutForm";
+import InjectedCheckoutForm from "./BillingCheckoutForm";
 
-//Wrap <Elements /> around your checkout form to group the set
+// Wrap <Elements /> around your checkout form to group the set
 // of Stripe Elements you're using together, so Stripe can pull
 // data from groups of Elements when tokenizing
 
-class MyStoreCheckout extends React.Component {
+class BillingStoreCheckout extends React.Component {
   render() {
     return (
       <Elements>
@@ -17,4 +17,4 @@ class MyStoreCheckout extends React.Component {
   }
 }
 
-export default MyStoreCheckout;
+export default BillingStoreCheckout;
