+++ front-end/src/components/Organisms/CalendarShiftUpdate.js
@@ -2,33 +2,38 @@ import React from "react";
 import { connect } from "react-redux";
 import moment from "moment";
 
-import { postShift } from "../../store/Shift/actions.js";
+import { updateShift, deleteShift } from "../../store/Shift/actions.js";
 
-import PostShiftTime from "./PostShiftTime.js";
+import PostShiftTime from "../Molecules/CalendarPostShiftTime.js";
 
 import {
-  Segment,
+  GridItemActiveShiftInner,
+  GridItemActiveShift,
+  ModalBackground,
+  ShiftConflictText,
+} from "../../styles/Calendar.js";
+import {
   Portal,
   Label,
   Header,
-  Form,
-  Input,
   Button,
+  Input,
   Divider,
+  Segment,
 } from "semantic-ui-react";
-import { GridItemShift, ModalBackground } from "../../styles/Calendar.js";
 
-class ScheduleShift extends React.Component {
+class ScheduleShiftUpdate extends React.Component {
   state = {
     open: false,
     clickX: 0,
     clickY: 0,
-    notes: undefined,
-    start_time: undefined,
-    end_time: undefined,
-    start_time24: undefined,
-    end_time24: undefined,
+    notes: this.props.notes,
+    start_time: moment(this.props.start).format("h:mm A"),
+    end_time: moment(this.props.end).format("h:mm A"),
+    start_time24: moment(this.props.start).format("H:mm"),
+    end_time24: moment(this.props.end).format("H:mm"),
     profile: this.props.profile,
+    id: this.props.id,
   };
 
   handleOpen = e => {
@@ -47,6 +52,10 @@ class ScheduleShift extends React.Component {
     this.setState({ open: false });
   };
 
+  handleDelete = () => {
+    this.props.deleteShift(this.props.id);
+  };
+
   submitHandler = event => {
     event.preventDefault();
 
@@ -60,7 +69,7 @@ class ScheduleShift extends React.Component {
         start_hour = this.state.start_time24.slice(0, 2);
         start_minutes = this.state.start_time24.slice(3, 5);
       }
-      const start_datetime = moment(this.props.date)
+      const start_datetime = moment(this.props.start)
         .second(0)
         .minute(start_minutes)
         .hour(start_hour)
@@ -80,13 +89,13 @@ class ScheduleShift extends React.Component {
       if (end_hour < start_hour) {
         // prettier-ignore
         end_datetime = moment(start_datetime)
-          .clone()
-          .add((23 - start_hour) + Number(end_hour), "hours")
-          .add((60 - start_minutes) + Number(end_minutes), "minutes")
-          .utc()
-          .format();
+            .clone()
+            .add((23 - start_hour) + Number(end_hour), "hours")
+            .add((60 - start_minutes) + Number(end_minutes), "minutes")
+            .utc()
+            .format();
       } else {
-        end_datetime = moment(this.props.date)
+        end_datetime = moment(this.props.end)
           .second(0)
           .minute(end_minutes)
           .hour(end_hour)
@@ -106,10 +115,11 @@ class ScheduleShift extends React.Component {
 
       const notes = this.state.notes ? this.state.notes : "";
 
-      this.props.postShift(
+      this.props.updateShift(
+        this.state.id,
+        profile,
         start_datetime,
         end_datetime,
-        profile,
         is_open,
         notes
       );
@@ -129,16 +139,24 @@ class ScheduleShift extends React.Component {
 
   render() {
     return (
-      <GridItemShift
-        key={`${this.props.row}-${this.props.column}`}
+      <GridItemActiveShift
         row={this.props.row}
         column={this.props.column}
-        background={this.props.row > 2 ? "none" : "hsl(104, 62.5%, 95%)"}
+        color={this.props.color}
       >
-        <div
-          style={{ width: "100%", height: "100%" }}
+        {this.props.conflict ? (
+          <ShiftConflictText>Availability Conflict</ShiftConflictText>
+        ) : null}
+        <GridItemActiveShiftInner
           onClick={this.handleOpen}
-        />
+          hue={this.props.hue}
+          start={this.props.startHour}
+          end={this.props.endHour}
+          conflict={this.props.conflict}
+        >
+          {moment(this.props.start).format("h:mm A")} -
+          {moment(this.props.end).format("h:mm A")}
+        </GridItemActiveShiftInner>
         <Portal
           open={this.state.open}
           onClose={this.handleClose}
@@ -152,7 +170,7 @@ class ScheduleShift extends React.Component {
                 left: `${
                   this.state.clickX > window.innerWidth - 142
                     ? this.state.clickX - 262
-                    : this.state.clickX
+                    : this.state.clickX - 50
                 }px`,
                 top: `${
                   this.state.clickY > window.innerHeight - 242
@@ -170,54 +188,75 @@ class ScheduleShift extends React.Component {
                 onClick={this.handleClose}
                 icon="close"
               />
-              <Form>
-                <Header as="h2" textAlign={"center"}>
-                  New Shift
-                </Header>
-                <Divider />
-                <PostShiftTime
-                  day={"Start Time"}
-                  data={"start"}
-                  start={this.state.start_time}
-                  inputChangeHandler={this.inputChangeHandler}
-                />
-                <PostShiftTime
-                  day={"End Time"}
-                  data={"end"}
-                  end={this.state.end_time}
-                  inputChangeHandler={this.inputChangeHandler}
-                />
-                <Input
-                  label={{ tag: true, content: "Note" }}
-                  labelPosition="left"
-                  style={{ width: "60%", margin: "8% 8%" }}
-                  value={this.state.notes}
-                  onChange={this.inputChangeHandler}
-                  name={"notes"}
-                />
-                <Button
-                  fluid
-                  size="large"
-                  color="teal"
-                  onClick={this.submitHandler}
-                >
-                  Submit
-                </Button>
-              </Form>
+              <Label
+                as="a"
+                color="red"
+                ribbon={"right"}
+                onClick={this.handleDelete}
+              >
+                Delete
+              </Label>
+              <Header as="h2" textAlign={"center"}>
+                Update Shift
+              </Header>
+              <Divider />
+              <PostShiftTime
+                day={"Start Time"}
+                start={this.state.start_time}
+                data={"start"}
+                inputChangeHandler={this.inputChangeHandler}
+              />
+              <PostShiftTime
+                day={"End Time"}
+                end={this.state.end_time}
+                data={"end"}
+                inputChangeHandler={this.inputChangeHandler}
+              />
+              <Input
+                label={{ tag: true, content: "Note" }}
+                labelPosition="left"
+                style={{ width: "70%", padding: "8% 8%" }}
+                value={this.state.notes}
+                onChange={this.inputChangeHandler}
+                name={"notes"}
+              />
+              <Button
+                fluid
+                size="large"
+                color="teal"
+                onClick={this.submitHandler}
+                onKeyPress={event => {
+                  if (event.key === "Enter") {
+                    this.submitHandler();
+                  }
+                }}
+              >
+                Submit
+              </Button>
             </Segment>
           </div>
         </Portal>
-      </GridItemShift>
+      </GridItemActiveShift>
     );
   }
 }
 
 const mapDispatchToProps = dispatch => {
   return {
-    postShift: (start_datetime, end_datetime, profile, is_open, notes) => {
+    updateShift: (
+      id,
+      profile,
+      start_datetime,
+      end_datetime,
+      is_open,
+      notes
+    ) => {
       return dispatch(
-        postShift(start_datetime, end_datetime, profile, is_open, notes)
+        updateShift(id, profile, start_datetime, end_datetime, is_open, notes)
       );
+    },
+    deleteShift: id => {
+      return dispatch(deleteShift(id));
     },
   };
 };
@@ -225,4 +264,4 @@ const mapDispatchToProps = dispatch => {
 export default connect(
   null,
   mapDispatchToProps
-)(ScheduleShift);
+)(ScheduleShiftUpdate);
