+++ front-end/src/components/Organisms/DashboardTimeOffApproved.js
@@ -57,6 +57,7 @@ class TimeOffApproved extends Component {
                     ) : null}
                   </Segment.Group>
                 );
+              else return null;
             })}
         </Segment>
       </TimeOffApprovedContainer>
