+++ front-end/src/components/Organisms/landingFooter.js
@@ -1,36 +1,40 @@
 import React, { Component } from "react";
-import { 
-    FooterStyles, 
-    LandingFoot
-    } from "../../styles/Landing.js";
-import Twitter from "../Molecules/TwitterIcon.js";
-import Facebook from "../Molecules/FacebookIcon.js";
 
-//import "../../styles/LandingRef.css";
-
+import { Header, Icon } from "semantic-ui-react";
+import { FooterStyles, LandingFoot } from "../../styles/Landing.js";
 
 class LandingFooter extends Component {
-    render(){
-        return (
-            <div>
-            <FooterStyles>
-                <LandingFoot>
-                <span className="copyright">Copyright 2018</span>
-                <div className="LandingLogo">
-                    <a href="http://twitter.com">
-                        <Twitter/>
-                    </a> 
-                </div> 
-                <div className="LandingLogo">
-                    <a href="http://facebook.com">
-                        <Facebook/>
-                    </a>
-                </div> 
-                </LandingFoot>
-            </FooterStyles>
+  render() {
+    return (
+      <div>
+        <FooterStyles>
+          <LandingFoot>
+            <span className="copyright">Copyright 2018</span>
+            <div className="LandingLogo">
+              <a href="http://twitter.com">
+                <Header as="h1">
+                  <Icon.Group size="small">
+                    <Icon name="twitter" />
+                    <Icon corner name="add" />
+                  </Icon.Group>
+                </Header>
+              </a>
             </div>
-        );
-    }
+            <div className="LandingLogo">
+              <a href="http://facebook.com">
+                <Header as="h1">
+                  <Icon.Group size="small">
+                    <Icon name="facebook" />
+                    <Icon corner name="add" />
+                  </Icon.Group>
+                </Header>
+              </a>
+            </div>
+          </LandingFoot>
+        </FooterStyles>
+      </div>
+    );
+  }
 }
 
 export default LandingFooter;
