+++ front-end/src/components/Pages/Billing.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 
 import { StripeProvider } from "react-stripe-elements";
 
-import MyStoreCheckout from "./test_pages/MyStoreCheckout";
+import BillingStoreCheckout from "../Organisms/BillingStoreCheckout";
 
 import {
   BillingContainer,
@@ -11,9 +11,9 @@ import {
 } from "../../styles/Billing.js";
 import { Segment } from "semantic-ui-react";
 
-//StripeProvider gives us access to the Stripe Object
-//i.e Stripe.createToken, stripe.elements() etc
-//App loads the stripe script asynchronously in CDM
+// StripeProvider gives us access to the Stripe Object
+// i.e Stripe.createToken, stripe.elements() etc
+// App loads the stripe script asynchronously in CDM
 
 class Billing extends Component {
   constructor(props) {
@@ -27,7 +27,7 @@ class Billing extends Component {
       });
     } else {
       document.querySelector("#stripe-js").addEventListener("load", () => {
-        //Create Stripe instance once Stripe.js loads
+        // Create Stripe instance once Stripe.js loads
         this.setState({
           stripe: window.Stripe(process.env.REACT_APP_publishable),
         });
@@ -43,7 +43,7 @@ class Billing extends Component {
           <Segment>
             {this.state.stripe ? (
               <StripeProvider stripe={this.state.stripe}>
-                <MyStoreCheckout />
+                <BillingStoreCheckout />
               </StripeProvider>
             ) : null}
           </Segment>
