+++ front-end/src/components/Pages/test_pages/Month.js
@@ -2,124 +2,128 @@ import React, { Component } from "react";
 import dateFns from "date-fns";
 // import {Link, Redirect } from "react-redux";
 // import Modal from "../test_pages/Modal.js";
-import "../../../styles/month.css";
- 
+import "./styles/month.css";
+
 class Month extends Component {
-    state = {
-        currentMonth: new Date(),
-        selectedDate: new Date()
+  state = {
+    currentMonth: new Date(),
+    selectedDate: new Date(),
+  };
+
+  // We create rendering functions
+
+  renderHeader() {
+    const dateFormat = "MMMM YYYY";
+
+    return (
+      <div className="header row flex-middle">
+        <div className="col col-start">
+          <div className="icon" onClick={this.prevMonth}>
+            chevron_left
+          </div>
+        </div>
+        <div className="col col-center">
+          <span>{dateFns.format(this.state.currentMonth, dateFormat)}</span>
+        </div>
+        <div className="col col-end">
+          <div className="icon" onClick={this.nextMonth}>
+            chevron_right
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  renderDays() {
+    const dateFormat = "dd";
+    const days = [];
+
+    let startDate = dateFns.startOfWeek(this.state.currentMonth);
+
+    for (let i = 0; i < 7; i++) {
+      days.push(
+        <div className="col col-center" key={i}>
+          {dateFns.format(dateFns.addDays(startDate, i), dateFormat)}
+        </div>
+      );
     }
+    return <div className="days row">{days}</div>;
+  }
 
-    // We create rendering functions
+  renderCells() {
+    const { currentMonth, selectedDate } = this.state;
+    const monthStart = dateFns.startOfMonth(currentMonth);
+    const monthEnd = dateFns.endOfMonth(monthStart);
+    const startDate = dateFns.startOfWeek(monthStart);
+    const endDate = dateFns.endOfWeek(monthEnd);
 
-    renderHeader() {
-        const dateFormat = "MMMM YYYY"
+    const dateFormat = "D";
+    const rows = [];
 
-        return( 
-            <div className="header row flex-middle">
-                <div className="col col-start">
-                    <div className="icon" onClick={this.prevMonth}>
-                        chevron_left
-                    </div>
-                </div>
-                <div className="col col-center">
-                    <span>
-                        {dateFns.format(this.state.currentMonth, dateFormat)}
-                    </span>
-                </div>
-                <div className="col col-end">
-                    <div className="icon" onClick={this.nextMonth}>
-                        chevron_right
-                    </div>
-                </div>
-            </div>
-        )
-    }
+    let days = [];
+    let day = startDate;
+    let formattedDate = "";
 
-    renderDays() {
-        const dateFormat = "dd";
-        const days = [];
-
-        let startDate = dateFns.startOfWeek(this.state.currentMonth);
-
-        for(let i = 0; i < 7; i++){
-            days.push(
-                <div className="col col-center" key={i}>
-                    {dateFns.format(dateFns.addDays(startDate, i), dateFormat )}
-                </div>
-            )
-        }
-        return <div className="days row">{days}</div>
-    }
-
-    renderCells() {
-        const { currentMonth, selectedDate } = this.state;
-        const monthStart = dateFns.startOfMonth(currentMonth);
-        const monthEnd = dateFns.endOfMonth(monthStart);
-        const startDate = dateFns.startOfWeek(monthStart);
-        const endDate = dateFns.endOfWeek(monthEnd);
-
-        const dateFormat = "D";
-        const rows = [];
-
-        let days = [];
-        let day = startDate;
-        let formattedDate = "";
-
-        while (day <= endDate){
-            for(let i = 0; i < 7; i++){
-                formattedDate = dateFns.format(day, dateFormat);
-                const cloneDay = day;
-                days.push(
-                    <div className={`col cell ${
-                        !dateFns.isSameMonth(day, monthStart)
-                            ? "disabled"
-                            : dateFns.isSameDay(day, selectedDate) ? "selected" : ""
-                        }`}
-                    key={day} onClick={() => this.onDateClick(dateFns.parse(cloneDay))}>
-                        <span className="number">{formattedDate}</span>
-                        {/* <Modal/> */}
-                    </div>
-                );
-                day = dateFns.addDays(day, 1);
-            }
-            rows.push(
-                <dv className="row" key={day}>
-                    {days}
-                </dv>
-            );
-            days = [];
-        }
-        return <div ClassName="body">{rows}</div>
-    }
-
-    // We will use these functions to control functionality
-
-    onDateClick = day => {
-        this.setState({ selectedDate: day });
-        console.log(this.state.selectedDate);
-    }
-
-
-
-    nextMonth = () => {
-        this.setState({ currentMonth: dateFns.addMonths(this.state.currentMonth, 1 )})
-    };
-
-    prevMonth = () => {
-        this.setState({ currentMonth: dateFns.subMonths(this.state.currentMonth, 1)})
-    };
-
-    render(){
-        return (
-            <div className="Month">
-                {this.renderHeader()}
-                {this.renderDays()}
-                {this.renderCells()}
-            </div>
+    while (day <= endDate) {
+      for (let i = 0; i < 7; i++) {
+        formattedDate = dateFns.format(day, dateFormat);
+        const cloneDay = day;
+        days.push(
+          <div
+            className={`col cell ${
+              !dateFns.isSameMonth(day, monthStart)
+                ? "disabled"
+                : dateFns.isSameDay(day, selectedDate)
+                  ? "selected"
+                  : ""
+            }`}
+            key={day}
+            onClick={() => this.onDateClick(dateFns.parse(cloneDay))}
+          >
+            <span className="number">{formattedDate}</span>
+            {/* <Modal/> */}
+          </div>
         );
+        day = dateFns.addDays(day, 1);
+      }
+      rows.push(
+        <dv className="row" key={day}>
+          {days}
+        </dv>
+      );
+      days = [];
     }
+    return <div ClassName="body">{rows}</div>;
+  }
+
+  // We will use these functions to control functionality
+
+  onDateClick = day => {
+    this.setState({ selectedDate: day });
+    console.log(this.state.selectedDate);
+  };
+
+  nextMonth = () => {
+    this.setState({
+      currentMonth: dateFns.addMonths(this.state.currentMonth, 1),
+    });
+  };
+
+  prevMonth = () => {
+    this.setState({
+      currentMonth: dateFns.subMonths(this.state.currentMonth, 1),
+    });
+  };
+
+  render() {
+    return (
+      <div className="Month">
+        {this.renderHeader()}
+        {this.renderDays()}
+        {this.renderCells()}
+      </div>
+    );
+  }
 }
 
 export default Month;
-
