+++ front-end/src/components/Pages/test_pages/month.css
@@ -1,4 +1,4 @@
-/* FONT IMPORTS */
+FONT IMPORTS
 
 @import url(https://fonts.googleapis.com/css?family=Open+Sans:300,400,700);
 @import url(https://fonts.googleapis.com/icon?family=Material+Icons);
