--- front-end/src/styles/Form.js
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-// STYLESHEET FOR ALL FORMS
